pyproject.toml
[project]
name = "pythonproject-news-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Rodion Marochkin",email = "rodionmarochkin32@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.2.0"
isort = "^6.0.1"
black = "^25.1.0"
mypy = "^1.15.0"


[tool.poetry.group.tests.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"


[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"

